requirements.txt
gym>=0.18.0
numpy>=1.16.6
pytorch-lightning>=1.2.4
scipy>=1.6.1